Moved /registration/submitmember to /members/submit
