+++ socrates/lib/middleware/redirectRuleForNewUser.js
@@ -2,7 +2,7 @@
 
 module.exports = function redirectRuleForNewUser(req, res, next) {
   function proceed() {
-    return (/\/members\/edit|\/registration\/submitmember|\/auth\/logout|clientscripts|stylesheets|img|fonts|checknickname|checkemail/).test(req.originalUrl);
+    return (/\/members\/edit|\/members\/submit|\/auth\/logout|clientscripts|stylesheets|img|fonts|checknickname|checkemail/).test(req.originalUrl);
   }
 
   if (req.user && !req.user.member && !proceed()) {
