+++ socrates/lib/registration/index.js
@@ -1,15 +1,9 @@
 'use strict';
 var moment = require('moment-timezone');
-var async = require('async');
-var _ = require('lodash');
 
 var beans = require('simple-configure').get('beans');
 var misc = beans.get('misc');
 var membersService = beans.get('membersService');
-var notifications = beans.get('notifications');
-var validation = beans.get('validation');
-var statusmessage = beans.get('statusmessage');
-var groupsAndMembersService = beans.get('groupsAndMembersService');
 var mailsenderService = beans.get('mailsenderService');
 
 var app = misc.expressAppIn(__dirname);
@@ -37,50 +31,6 @@ app.get('/participate', function (req, res) {
 app.post('/participate', function (req, res) {
   console.log(req.body);
   res.redirect('participate');
-});
-
-app.post('/submitmember', function (req, res, next) {
-  function memberSubmitted(req, res, next) {
-    function notifyNewMemberRegistration(member, subscriptions) {
-      // must be done here, not in Service to avoid circular deps
-      return notifications.newMemberRegistered(member, subscriptions);
-    }
-
-    return groupsAndMembersService.updateAndSaveSubmittedMemberWithoutSubscriptions(req.user, req.body, res.locals.accessrights, notifyNewMemberRegistration, function (err, nickname) {
-      if (err) { return next(err); }
-      if (nickname) {
-        statusmessage.successMessage('message.title.save_successful', 'message.content.members.saved').putIntoSession(req);
-        return res.redirect('/');
-      }
-      return res.redirect('/');
-    });
-  }
-
-  async.parallel(
-    [
-      function (callback) {
-        // we need this helper function (in order to have a closure?!)
-        var validityChecker = function (nickname, callback) { membersService.isValidNickname(nickname, callback); };
-        validation.checkValidity(req.body.previousNickname, req.body.nickname, validityChecker, 'validation.nickname_not_available', callback);
-      },
-      function (callback) {
-        // we need this helper function (in order to have a closure?!)
-        var validityChecker = function (email, callback) { membersService.isValidEmail(email, callback); };
-        validation.checkValidity(req.body.previousEmail, req.body.email, validityChecker, 'validation.duplicate_email', callback);
-      },
-      function (callback) {
-        return callback(null, validation.isValidForSoCraTesMember(req.body));
-      }
-    ],
-    function (err, errorMessages) {
-      var realErrors = _.filter(_.flatten(errorMessages), function (message) { return !!message; });
-      if (realErrors.length === 0) {
-        return memberSubmitted(req, res, next);
-      }
-      return res.render('../../../views/errorPages/validationError', {errors: realErrors});
-    }
-  );
-
 });
 
 app.get('/resign', function (req, res) {
